+++ src/components/Container.module.scss
@@ -1,0 +1,5 @@
+@import '../styles/shared/mixins';
+
+.container {
+  @include box-padding;
+}
