+++ src/components/Container.tsx
@@ -1,5 +1,12 @@
 import * as React from 'react'
+import ContainerStyles from './Container.module.scss'
 
-const Container: React.FC = ({ children }) => <>{children}</>
+interface ContainerProps {
+  className: string
+}
+
+const Container: React.FC<ContainerProps> = ({ className, children }) => (
+  <section className={`${className} ${ContainerStyles.container}`}>{children}</section>
+)
 
 export default Container
