+++ src/components/blog-highllights/BlogHighlightSection.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react'
 import BlogHighlightSectionStyles from './BlogHighlightSection.module.css'
 import { ExerptList } from './exerpt-list/ExerptList'
+import Container from '../Container'
 
 const BlogHighlightSection: React.FC<{}> = () => {
   return (
-    <section className={BlogHighlightSectionStyles.container}>
+    <Container className={BlogHighlightSectionStyles.container}>
       <h1>Catch Up On The Latest</h1>
       <ExerptList />
-    </section>
+    </Container>
   )
 }
