--- src/components/card/card-body/CardBody.module.css
@@ -1,7 +1,0 @@
-.card {
-  background-color: rgba(255, 255, 255, 0.05);
-  display: flex;
-  flex-flow: column nowrap;
-  margin: 16px;
-  padding: 16px;
-}
