+++ src/components/card/card-body/CardBody.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import CardBodyStyles from './CardBody.module.css'
+import CardBodyStyles from './CardBody.module.scss'
 
 const CardBody: React.FC = ({ children }) => <article className={CardBodyStyles.card}>{children}</article>
