--- src/pages/404.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react'
-import { Link } from 'gatsby'
-
-import Page from '../components/Page'
-import Container from '../components/Container'
-import IndexLayout from '../layouts'
-
-const NotFoundPage = () => (
-  <IndexLayout>
-    <Page>
-      <Container>
-        <h1>404: Page not found.</h1>
-        <p>
-          You've hit the void. <Link to="/">Go back.</Link>
-        </p>
-      </Container>
-    </Page>
-  </IndexLayout>
-)
-
-export default NotFoundPage
