+++ src/styles/global/_typeography.scss
@@ -48,7 +48,7 @@ h4,
   line-height: 20px;
 }
 
-.text-lead {
+.text-style-muted {
   font-family: $font-family-body;
   font-style: normal;
   font-weight: 300;
@@ -57,7 +57,7 @@ h4,
   letter-spacing: 0.02em;
 }
 
-.text-lead {
+.text-style-lead {
   font-family: $font-family-body;
   font-style: normal;
   font-weight: 300;
@@ -66,7 +66,7 @@ h4,
   letter-spacing: 0.04em;
 }
 
-.text-label {
+.text-style-label {
   font-family: $font-family-body;
   font-style: normal;
   font-weight: bold;
@@ -75,10 +75,22 @@ h4,
   letter-spacing: 0.04em;
 }
 
-.text-code {
+.text-style-code {
   font-family: $font-family-code;
   font-style: normal;
   font-weight: normal;
   font-size: 1rem;
   line-height: 21px;
+}
+
+.text-color-primary {
+  color: $text-color-primary;
+}
+
+.text-color-secondary {
+  color: $text-color-secondary;
+}
+
+.text-color-disabled {
+  color: $text-color-disabled;
 }
