+++ src/styles/shared/_variables.scss
@@ -9,8 +9,8 @@ $background-primary: #121212;
 $background-accent: #121f20;
 $color-modify-lighten-1: rgba(255, 255, 255, 0.05);
 $color-modify-lighten-2: rgba(255, 255, 255, 0.22);
-$text-emphasis-high: rgba(255, 255, 255, 0.98);
-$text-emphisis-medium: rgba(255, 255, 255, 0.6);
-$text-emphisis-disabled: rgba(255, 255, 255, 0.38);
+$text-color-primary: rgba(255, 255, 255, 0.98);
+$text-color-secondary: rgba(255, 255, 255, 0.6);
+$text-color-disabled: rgba(255, 255, 255, 0.38);
 $accent-primary: #70f3ff;
 $accent-secondary: #4a99a0;
